pyproject.toml
[project]
name = "pdf-filler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.5.4",
    "pymupdf>=1.26.3",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.5",
]
